# api-gateway/docs/swagger.yaml
definitions:
...
  hotelproto.Hotel:
    properties:
      address:
        type: string
      hotelId:
        type: string
      location:
        type: string
      name:
        type: string
      rating:
        type: integer
    type: object
  hotelproto.HotelRequest:
    properties:
      address:
        type: string
      location:
        type: string
      name:
        type: string
      rating:
        type: integer
    type: object
  hotelproto.HotelRes:
    properties:
      message:
        type: string
    type: object
  hotelproto.HotelResponse:
    properties:
      hotelId:
        type: string
    type: object
  hotelproto.ListHotels:
    properties:
      hotel:
        items:
          $ref: '#/definitions/hotelproto.Hotel'
        type: array
    type: object
  hotelproto.ListRooms:
    properties:
      room:
        items:
          $ref: '#/definitions/hotelproto.Room'
        type: array
    type: object
  hotelproto.Room:
    properties:
      availability:
        type: boolean
      hotelId:
        type: string
      pricePerNight:
        type: integer
      roomId:
        type: string
      roomType:
        type: string
    type: object
  hotelproto.RoomRequest:
    properties:
      availability:
        type: boolean
      hotelId:
        type: string
      pricePerNight:
        type: integer
      roomType:
        type: string
    type: object
  hotelproto.RoomRes:
    properties:
      message:
        type: string
    type: object
  hotelproto.RoomResponse:
    properties:
      roomId:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample server for a restaurant reservation system.
  title: Hotel Booking System
  version: "1.0"
paths:
...
  /hotels:
    get:
      description: Get all hotels
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hotelproto.ListHotels'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Get all hotels
      tags:
      - hotel
    post:
      consumes:
      - application/json
      description: Create a new hotel
      parameters:
      - description: Hotel request body
        in: body
        name: hotel
        required: true
        schema:
          $ref: '#/definitions/hotelproto.HotelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hotelproto.HotelResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Create a new hotel
      tags:
      - hotel
  /hotels/{id}:
    delete:
      description: Delete hotel by ID
      parameters:
      - description: Hotel ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hotelproto.HotelRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Delete hotel by ID
      tags:
      - hotel
    get:
      consumes:
      - application/json
      description: Get hotel by ID
      parameters:
      - description: Hotel ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hotelproto.Hotel'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Get hotel by ID
      tags:
      - hotel
    put:
      consumes:
      - application/json
      description: Update hotel by ID
      parameters:
      - description: Hotel ID
        in: path
        name: id
        required: true
        type: string
      - description: Hotel update request body
        in: body
        name: hotel
        required: true
        schema:
          $ref: '#/definitions/hotelproto.Hotel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hotelproto.HotelRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Update hotel by ID
      tags:
      - hotel
...
  /rooms:
    get:
      description: Get all rooms
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hotelproto.ListRooms'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Get all rooms
      tags:
      - room
    post:
      consumes:
      - application/json
      description: Create a new room
      parameters:
      - description: Room request body
        in: body
        name: room
        required: true
        schema:
          $ref: '#/definitions/hotelproto.RoomRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hotelproto.RoomResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Create a new room
      tags:
      - room
  /rooms/{id}:
    delete:
      description: Delete room by ID
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hotelproto.RoomRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Delete room by ID
      tags:
      - room
    get:
      description: Get room by ID
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hotelproto.Room'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Get room by ID
      tags:
      - room
    put:
      consumes:
      - application/json
      description: Update room by ID
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: string
      - description: Room update request body
        in: body
        name: room
        required: true
        schema:
          $ref: '#/definitions/hotelproto.Room'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hotelproto.RoomRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Update room by ID
      tags:
      - room
...
securityDefinitions:
  Bearer:
    description: Enter the token in the format `Bearer {token}`
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"